Update README.md
We chose these numbers to assure that each additional step we take is going to cost and difference between the cells of the grid is noticeable. The goal has reward zero. In the following, we have tested different SARSA methods individually based on the different alpha parameter. Finally, we will compare all of them together assuming that the alpha is equal to 1.